Testing_Shaders/Created Combo Score script
Attempted to work on shadow script (not working)

Created a combo score script (only created have not worked on it)

+++ Based Runner Game/Assets/Scripts/DataPersistence/Data/GameData.cs
@@ -17,7 +17,7 @@ public class GameData
 
     public SerializableDictionary<int, bool> Extras_;
     public SerializableDictionary<int, bool> Levels_Unlocked;
-    public SerializableDictionary<int, int> Level_HighScores;
+    public SerializableDictionary<int, float> Level_HighScores;
 
 
     public GameData()
@@ -32,6 +32,6 @@ public class GameData
 
         Levels_Unlocked = new SerializableDictionary<int, bool>();
 
-        Level_HighScores = new SerializableDictionary<int, int>();
+        Level_HighScores = new SerializableDictionary<int, float>();
     }
 }
